+++ Hardware/Motor.h
@@ -2,7 +2,7 @@
 #define __MOTOR_H
 
 void Motor_Init(void);
-void Motor1_SetSpeed(int8_t Speed);
-void Motor2_SetSpeed(int8_t Speed);
+void Motor1_SetSpeed(int Speed);
+void Motor2_SetSpeed(int Speed);
 
 #endif
